- moved contextbuilder, context and expectedsetcall out to their own files - corrected x86 build with get path from running path test, cmd.exe runs from SysWOW64 in x86 rather than system32
